data_collection_mechanisms/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via starlette
asgiref==3.8.1
    # via django
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
django==4.2
    # via -r requirements.in
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
pydantic==2.10.4
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via django
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
